requirements.txt
py-cord
Pillow
